feat : Change item from struct type to class type to make it conform to NSCoding 	   Add additional code related to NSCoding, NSKeyedArchiver to persist data  	   Add the code in AppDelegate file that fetch seed list file on the disk

+++ ShoppingList/ShoppingList/ItemAddViewController.swift
@@ -45,7 +45,7 @@ class ItemAddViewController : UIViewController {
         // appDelegate.items.append(newItem)
         
         //
-        let newItem = Item(price: Double(priceTextField.text!)!, name: nameTextField.text!)
+        let newItem = Item(price: Float(priceTextField.text!)!, name: nameTextField.text!)
         delegate!.addItemInList(self, item: newItem)
         
         self.dismissViewControllerAnimated(true, completion: nil)
